// functions/src/interfaces.ts
export interface InvoicePayload {
  email?: string;
  uid?: string;
  items: [OrderItem];
  daysUntilDue?: number;
  default_tax_rates?: string[];
  transfer_data?: {
    destination: string;
    amount?: number;
  };
}

export interface OrderItem {
  amount: number;
  currency: string;
  quantity?: number;
  description: string;
  tax_rates?: string[];
}